[refactor] 디렉토리 구성 수정, 파일명 일부 수정
