+++ src/main/java/Opensource_SW_Project/Project/service/ChatgptApiService/ChatgptApiCommandServiceImpl.java
@@ -7,9 +7,9 @@ import Opensource_SW_Project.Project.domain.enums.Category;
 import Opensource_SW_Project.Project.repository.DetailedTalkRepository;
 import Opensource_SW_Project.Project.repository.TalkRepository;
 import Opensource_SW_Project.Project.repository.UserRepository;
-import Opensource_SW_Project.Project.web.dto.ChatGPTRequestDTO;
-import Opensource_SW_Project.Project.web.dto.ChatGPTResponseDTO;
-import Opensource_SW_Project.Project.web.dto.TalkRequestDTO;
+import Opensource_SW_Project.Project.web.dto.ChatGPT.ChatGPTRequestDTO;
+import Opensource_SW_Project.Project.web.dto.ChatGPT.ChatGPTResponseDTO;
+import Opensource_SW_Project.Project.web.dto.Talk.TalkRequestDTO;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
